# docs/openapi.yaml
openapi: 3.0.3
info:
  title: SaaS FinOps Analytics Platform API
  version: 1.0.0
  description: |
    The SaaS FinOps Analytics Platform provides multi-tenant cloud cost management,
    AI-powered database analytics, and subscription billing capabilities.

    ## Authentication

    The API supports two authentication methods:

    ### JWT Bearer Token
    Obtain a token via `POST /auth/login`. Include it in the `Authorization` header:
    ```
    Authorization: Bearer <access_token>
    ```
    Access tokens expire after **15 minutes**. Use `POST /auth/refresh` to obtain a new one.

    ### API Key
    Generate an API key via `POST /auth/api-keys`. Include it in the `X-API-Key` header:
    ```
    X-API-Key: <your_api_key>
    ```

    ## Rate Limiting

    Rate limits are enforced per user per minute based on subscription tier:

    | Plan       | Requests/min |
    |------------|-------------|
    | Free       | 100         |
    | Base       | 500         |
    | Pro        | 2,000       |
    | Enterprise | 10,000      |

    When the limit is exceeded, the API returns `429 Too Many Requests` with a `Retry-After` header.

  contact:
    name: Platform Support
    email: support@finops-platform.example.com
  license:
    name: Proprietary
servers:
  - url: http://localhost:8080
    description: Local development (API Gateway)
  - url: http://localhost:8081
    description: Auth Service (direct)
  - url: http://localhost:8082
    description: Billing Service (direct)
...
security:
  - BearerAuth: []
  - ApiKeyAuth: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token obtained from POST /auth/login. Expires in 15 minutes.
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key generated via POST /auth/api-keys.
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
          example: "invalid credentials"
        message:
          type: string
          example: "The provided email or password is incorrect"
      required: [error]
...
    RateLimitError:
      type: object
      properties:
        error:
          type: string
          example: "rate limit exceeded"
        retry_after:
          type: integer
          example: 30
          description: Seconds until the rate limit resets
...
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
          example: "user@example.com"
        password:
          type: string
          example: "SecurePass123!"
    LoginResponse:
      type: object
      properties:
        access_token:
          type: string
          example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
        refresh_token:
          type: string
          example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
        expires_in:
          type: integer
          example: 900
          description: Access token lifetime in seconds
...
    CreateApiKeyRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
          example: "Production API Key"
        expires_in_days:
          type: integer
          example: 365
          default: 365
    CreateApiKeyResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: "550e8400-e29b-41d4-a716-446655440001"
        api_key:
          type: string
          example: "sk_live_abcdefghijklmnopqrstuvwxyz123456"
          description: "32-character key. Shown only once - save it securely."
        warning:
          type: string
          example: "Save this key, it won't be shown again"
    ApiKey:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
          example: "Production API Key"
        last_used_at:
          type: string
          format: date-time
          nullable: true
        expires_at:
          type: string
          format: date-time
        created_at:
          type: string
          format: date-time
...
paths:
...
  /auth/login:
    post:
      tags: [auth]
      summary: Authenticate user and obtain tokens
      description: |
        Authenticates a user with email and password. Returns a short-lived access token
        (15 minutes) and a long-lived refresh token (7 days).

        **Rate limit:** 100 req/min (Free), 500 (Base), 2000 (Pro), 10000 (Enterprise)
      operationId: loginUser
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
            example:
              email: "alice@acme.com"
              password: "SecurePass123!"
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
              example:
                access_token: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VyX2lkIjoiNTUwZTg0MDAtZTI5Yi00MWQ0LWE3MTYtNDQ2NjU1NDQwMDAwIiwiYWNjb3VudF9pZCI6IjY2MGU4NDAwLWUyOWItNDFkNC1hNzE2LTQ0NjY1NTQ0MDAwMCIsInJvbGVzIjpbImFjY291bnRfb3duZXIiXSwiZXhwIjoxNzA2NzQ0MDAwfQ.signature"
                refresh_token: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VyX2lkIjoiNTUwZTg0MDAtZTI5Yi00MWQ0LWE3MTYtNDQ2NjU1NDQwMDAwIiwiZXhwIjoxNzA3MzQ4MDAwfQ.signature"
                expires_in: 900
        '400':
          description: Invalid request body
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Invalid credentials or unverified email
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                error: "invalid credentials"
        '429':
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RateLimitError'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /auth/api-keys:
    post:
      tags: [api-keys]
      summary: Create a new API key
      description: |
        Generates a cryptographically secure 32-character API key.
        **The key is shown only once** - store it securely immediately after creation.
        Keys expire after 365 days by default.
      operationId: createApiKey
      security:
        - BearerAuth: []
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateApiKeyRequest'
            example:
              name: "CI/CD Pipeline Key"
              expires_in_days: 365
      responses:
        '201':
          description: API key created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateApiKeyResponse'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '429':
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RateLimitError'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    get:
      tags: [api-keys]
      summary: List API keys
      description: Returns all active API keys for the authenticated user (key values are not returned).
      operationId: listApiKeys
      security:
        - BearerAuth: []
        - ApiKeyAuth: []
      responses:
        '200':
          description: List of API keys
          content:
            application/json:
              schema:
                type: object
                properties:
                  api_keys:
                    type: array
                    items:
                      $ref: '#/components/schemas/ApiKey'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...